Move config repository and split interfaces with impl
