+++ app/ui/latest/src/main/java/com/mudita/mail/repository/auth/config/AuthConfigRepository.kt
@@ -1,0 +1,8 @@
+package com.mudita.mail.repository.auth.config
+
+import com.mudita.mail.service.auth.config.AuthConfig
+
+interface AuthConfigRepository {
+
+    fun getAuthConfig(predicate: (AuthConfig) -> Boolean): AuthConfig?
+}
