+++ app/ui/latest/src/main/java/com/mudita/mail/repository/auth/config/PredefinedAuthConfigRepository.kt
@@ -1,14 +1,9 @@
-package com.mudita.mail.repository.providers.authConfig
+package com.mudita.mail.repository.auth.config
 
 import android.content.Context
 import com.mudita.mail.R
-import com.mudita.mail.service.authConfig.AuthConfig
-import com.mudita.mail.service.authConfig.ResponseType
-
-interface AuthConfigRepository {
-
-    fun getAuthConfig(predicate: (AuthConfig) -> Boolean): AuthConfig?
-}
+import com.mudita.mail.service.auth.config.AuthConfig
+import com.mudita.mail.service.auth.config.ResponseType
 
 class PredefinedAuthConfigRepository(context: Context) : AuthConfigRepository {
 
@@ -18,7 +13,9 @@ class PredefinedAuthConfigRepository(context: Context) : AuthConfigRepository {
             clientId = context.getString(R.string.google_client_id),
             redirectUrl = context.getString(R.string.google_redirect_url),
             scope = "https://mail.google.com/",
-            responseType = ResponseType.CODE
+            responseType = ResponseType.CODE,
+            authEndpoint = "https://accounts.google.com/o/oauth2/auth",
+            tokenEndpoint = "https://oauth2.googleapis.com/token"
         )
     )
